requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
anyio==4.1.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
boto3==1.34.1
botocore==1.34.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
decorator==5.1.1
distro==1.8.0
duckdb==0.8.1
exceptiongroup==1.2.0
executing==2.0.1
Faker==19.13.0
fonttools==4.46.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-play-scraper==1.2.4
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0
langchain-experimental==0.0.47
langsmith==0.0.70
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.9
packaging==23.2
pandas==1.5.3
pandasai==1.5.7
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
s3transfer==0.9.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.50
stack-data==0.6.3
streamlit==1.29.0
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.25.4
validators==0.22.0
watermark==2.4.3
wcwidth==0.2.12
yarl==1.9.4
zipp==3.17.0
